refactor: move gpiod_compat shim to device package
