+++ src/yoyopod/ui/input/adapters/gpiod_buttons.py
@@ -23,7 +23,7 @@ from loguru import logger
 
 from yoyopod.ui.input.hal import InputAction, InputHAL
 
-from yoyopod.ui.gpiod_compat import (
+from yoyopod.device.gpiod_compat import (
     HAS_GPIOD,
     get_event_fd,
     open_chip,
